A whole-number field holds a long
FIX writes int, Length, NumInGroup, SeqNum, TagNum and DayOfMonth as digits of any length,
and the model held them in a BigInteger so as to lose nothing. Nothing on a wire comes near
nineteen digits: the codes are one to three, the counters and lengths fit an int, a sequence
number a long many times over. What does not fit a long is not a value, and sets IsValid to
false, as a value that does not fit a decimal already did.

A long compares by an instruction and is a constant to C#, so the 140 whole-number code
checks are written as patterns, which Roslyn compiles to a range test or a jump table. The
readers add up to eighteen digits unchecked and more of them checked at every step, below
zero, so that long.MinValue is read. System.Numerics leaves the package.

+++ src/DotGram.Finance/Fix/FixParser.Messages.cs
@@ -2,7 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Numerics;
 
 namespace DotGram.Finance.Fix;
 
@@ -144,7 +143,7 @@ public static partial class FixParser
 	// is not kept, and a FIX Length is plain digits, so the two agree wherever the input is valid.
 	static FixField LengthField(int tag, int count, FixCustomFields? custom)
 	{
-		(bool, BigInteger) value = (true, count);
+		(bool, long) value = (true, count);
 
 		return tag switch
 		{
